Move common structures to com module
