+++ src/asm.rs
@@ -1,11 +1,4 @@
 //! Submodules for parsing and encoding HACK assembly programs.
 
-pub mod addr;
-pub mod comp;
-pub mod dest;
-pub mod inst;
-pub mod jump;
-pub mod label;
 pub mod parser;
 pub mod prog;
-pub mod symbol;
