+++ src/com.rs
@@ -1,0 +1,9 @@
+//! Submodules for common structures.
+
+pub mod addr;
+pub mod comp;
+pub mod dest;
+pub mod inst;
+pub mod jump;
+pub mod label;
+pub mod symbol;
