+++ src/lib.rs
@@ -7,5 +7,6 @@
 //! assembler translates HACK assembly code to HACK machine code.
 
 pub mod asm;
+pub mod com;
 pub mod dis;
 pub mod utils;
